Refactored authentication system
