+++ tooling_project/templates/main.html
@@ -63,8 +63,8 @@
             </li>
           </ul>
           <ul class="nav navbar-nav pull-right">
-            <li><a href="{{ url_for('main.register') }}">Register</a></li>
-            <li><a href="{{ url_for('main.login') }}">Login</a></li>
+            <li><a href="{{ url_for('auth.register') }}">Register</a></li>
+            <li><a href="{{ url_for('auth.login') }}">Login</a></li>
           </ul>
         </div><!--/.nav-collapse -->
       </div>
